requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
openai
chromadb
pypdf
tiktoken
numpy
pandas
unstructured